refactor passes switch recovery tests
